\name{popRange-package}
\alias{popRange-package}
\alias{popRange}
\docType{package}
\title{
popRange: A spatially and temporally explicit forward genetic simulator
}
\description{
The R package is a wrapper for a python program that simulates stochastic, complex population genetic models.
}
\details{
\tabular{ll}{
Package: \tab popRange\cr
Type: \tab Package\cr
Version: \tab 1.1.3\cr
Date: \tab 2014-10-08\cr
License: \tab MIT | share/LICENSE\cr
}
This R package consists of one main function, popRangeSim, which is a wrapper for the python code that simulates the populations.
}
\author{
\tabular{ll}{
Author: \tab Kimberly F. McManus\cr
Maintainer: \tab Kimberly F. McManus <kfm@stanford.edu>\cr
}
}
\references{ 
	Please see http://stanford.edu/~kfm/html/research.html for extended manual and examples.
	 }

